refactor: Rename module to 'boundary'
