+++ arcade-boundaries/src/main/kotlin/net/casual/arcade/boundary/renderer/BoundaryRenderer.kt
@@ -2,12 +2,12 @@
  * Copyright (c) 2025 senseiwells
  * Licensed under the MIT License. See LICENSE file in the project root for details.
  */
-package net.casual.arcade.border.renderer
+package net.casual.arcade.boundary.renderer
 
 import com.mojang.serialization.Codec
 import com.mojang.serialization.MapCodec
-import net.casual.arcade.border.shape.BoundaryShape
-import net.casual.arcade.border.utils.BoundaryRegistries
+import net.casual.arcade.boundary.shape.BoundaryShape
+import net.casual.arcade.boundary.utils.BoundaryRegistries
 import net.casual.arcade.utils.codec.CodecProvider.Companion.register
 import net.minecraft.core.Registry
 import net.minecraft.network.protocol.Packet
