+++ arcade-boundaries/src/main/kotlin/net/casual/arcade/boundary/renderer/options/AxisAlignedModelCubeShaderRenderOptions.kt
@@ -2,9 +2,9 @@
  * Copyright (c) 2025 senseiwells
  * Licensed under the MIT License. See LICENSE file in the project root for details.
  */
-package net.casual.arcade.border.renderer.options
+package net.casual.arcade.boundary.renderer.options
 
-import net.casual.arcade.border.shape.BoundaryShape
+import net.casual.arcade.boundary.shape.BoundaryShape
 import net.casual.arcade.utils.ArcadeUtils
 import net.minecraft.core.Direction
 import net.minecraft.core.component.DataComponents
@@ -12,7 +12,6 @@ import net.minecraft.util.Brightness
 import net.minecraft.world.item.ItemStack
 import net.minecraft.world.item.Items
 import net.minecraft.world.item.component.DyedItemColor
-import kotlin.math.ceil
 
 public class AxisAlignedModelCubeShaderRenderOptions: AxisAlignedModelRenderOptions {
     private val stationary = ArcadeUtils.id("boundary/cube_stationary")
