+++ arcade-boundaries/src/main/kotlin/net/casual/arcade/boundary/renderer/options/AxisAlignedModelRenderOptions.kt
@@ -2,12 +2,12 @@
  * Copyright (c) 2025 senseiwells
  * Licensed under the MIT License. See LICENSE file in the project root for details.
  */
-package net.casual.arcade.border.renderer.options
+package net.casual.arcade.boundary.renderer.options
 
 import com.mojang.serialization.Codec
-import net.casual.arcade.border.renderer.AxisAlignedDisplayBoundaryRenderer
-import net.casual.arcade.border.shape.BoundaryShape
-import net.casual.arcade.border.utils.BoundaryRegistries
+import net.casual.arcade.boundary.renderer.AxisAlignedDisplayBoundaryRenderer
+import net.casual.arcade.boundary.shape.BoundaryShape
+import net.casual.arcade.boundary.utils.BoundaryRegistries
 import net.casual.arcade.resources.ArcadeResourcePacks
 import net.casual.arcade.utils.ArcadeUtils
 import net.minecraft.core.Direction
