+++ arcade-boundaries/src/main/kotlin/net/casual/arcade/boundary/utils/BoundaryUtils.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2025 senseiwells
  * Licensed under the MIT License. See LICENSE file in the project root for details.
  */
-package net.casual.arcade.border.utils
+package net.casual.arcade.boundary.utils
 
 import net.minecraft.network.protocol.game.*
 import net.minecraft.world.level.border.WorldBorder
