+++ src/testmod/kotlin/net/casual/arcade/test/command/LevelBoundaryCommand.kt
@@ -3,11 +3,11 @@ package net.casual.arcade.test.command
 import com.mojang.brigadier.arguments.IntegerArgumentType
 import com.mojang.brigadier.builder.LiteralArgumentBuilder
 import com.mojang.brigadier.context.CommandContext
-import net.casual.arcade.border.LevelBoundary
-import net.casual.arcade.border.extension.LevelBoundaryExtension.Companion.levelBoundary
-import net.casual.arcade.border.renderer.AxisAlignedDisplayBoundaryRenderer
-import net.casual.arcade.border.renderer.options.AxisAlignedModelRenderOptions
-import net.casual.arcade.border.shape.AxisAlignedBoundaryShape
+import net.casual.arcade.boundary.LevelBoundary
+import net.casual.arcade.boundary.extension.LevelBoundaryExtension.Companion.levelBoundary
+import net.casual.arcade.boundary.renderer.AxisAlignedDisplayBoundaryRenderer
+import net.casual.arcade.boundary.renderer.options.AxisAlignedModelRenderOptions
+import net.casual.arcade.boundary.shape.AxisAlignedBoundaryShape
 import net.casual.arcade.commands.*
 import net.casual.arcade.utils.TimeUtils.Ticks
 import net.minecraft.commands.CommandBuildContext
